+++ src/websocket_server/enums/ShotStatusType.ts
@@ -1,0 +1,5 @@
+export const enum ShotStatusType {
+  MISS = 'miss',
+  KILLED = 'killed',
+  SHOT = 'shot',
+}
