+++ src/websocket_server/enums/index.ts
@@ -1,1 +1,2 @@
 export * from './EventType';
+export * from './ShotStatusType';
