+++ src/websocket_server/handlers/index.ts
@@ -1,2 +1,8 @@
 export * from './updateRoom';
 export * from './updateWinners';
+export * from './regestrationUser';
+export * from './addUserToRoom';
+export * from './addShips';
+export * from './universalAttack';
+export * from './userAttack';
+export * from './randomAttack';
