+++ src/websocket_server/handlers/updateRoom.ts
@@ -1,10 +1,10 @@
-import { WebSocket } from 'ws';
 import { getRoomsWithOneUser } from '../db';
 import { EventType } from '../enums';
+import { ISocket } from '../models';
 
-export const updateRoom = (sockets: WebSocket[]) => {
-  sockets.forEach((socket) =>
-    socket.send(
+export const updateRoom = (sockets: ISocket[]) => {
+  sockets.forEach(({ ws }) =>
+    ws.send(
       JSON.stringify({
         type: EventType.UPDATE_ROOM,
         data: JSON.stringify(getRoomsWithOneUser()),
