+++ src/websocket_server/handlers/updateWinners.ts
@@ -1,10 +1,10 @@
-import { WebSocket } from 'ws';
 import { getWinners } from '../db';
 import { EventType } from '../enums';
+import { ISocket } from '../models';
 
-export const updateWinners = (sockets: WebSocket[]) => {
-  sockets.forEach((socket) =>
-    socket.send(
+export const updateWinners = (sockets: ISocket[]) => {
+  sockets.forEach(({ ws }) =>
+    ws.send(
       JSON.stringify({
         type: EventType.UPDATE_WINNERS,
         data: JSON.stringify(getWinners()),
