+++ src/websocket_server/models/IGame.ts
@@ -1,11 +1,3 @@
-export const enum ShotStatusType {
-  MISS = 'miss',
-  KILLED = 'killed',
-  SHOT = 'shot',
-}
-
-type TypeOfSizeShip = 'small' | 'medium' | 'large' | 'huge';
-
 export interface IPosition {
   x: number;
   y: number;
@@ -20,7 +12,6 @@ export interface IShip {
   direction: boolean;
   length: number;
   position: IPosition;
-  type: TypeOfSizeShip;
   decks: IDeck[];
   boundaryСells: IPosition[];
 }
