+++ src/websocket_server/models/ISocket.ts
@@ -1,0 +1,6 @@
+import { WebSocket } from 'ws';
+
+export interface ISocket {
+  ws: WebSocket;
+  userIndex: number;
+}
