+++ src/websocket_server/models/index.ts
@@ -2,3 +2,4 @@ export * from './IUser';
 export * from './IGame';
 export * from './IRoom';
 export * from './IWinner';
+export * from './ISocket';
