+++ src/websocket_server/utils/generateRandomPosition.ts
@@ -1,0 +1,1 @@
+export const generateRandomPosition = () => Math.floor(Math.random() * 10);
