--- todo.txt
@@ -1,7 +1,0 @@
-1) определиться с тем, что сначала в базе меняем а потом на фронт или наоборот?
-2) рефкторинг
-3) как правильно обновлять стор?
-4) рандомный выстрел обрабатывать
-5) игра с ботом
-6) дисконнект правильный
-// TODO: можно turnIndexUser сделать рандомный. Или посмотреть в тз как вообще должно быть
